requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_openai==0.1.7
polars==0.20.26
python-dotenv==1.0.1
Requests==2.31.0
